refactor(formio): cut the trigger sample path down to what it uses
Dropping the polling triggers left the cursor unreachable: the only
caller is the webhook triggers' test(), and pollingHelper.test passes
lastFetchEpochMS 0, which always took the sample branch. The store
read, the high-water mark, the __gt filter and the oldest-first paging
could not run, so they went, and the module is renamed for what it
actually does now. The two helpers move to named parameters.

The test button behaves as before apart from fetching 25 rows rather
than 100, which pollingHelper.test slices to 5 either way.

Tests drop to 31, and the survivors are mutation-checked: sorting
oldest-first fails two, dropping the never-edited filter fails one.
